+++ Assets/Scripts/Common/GameStartMessage.cs
@@ -1,26 +1,20 @@
 ﻿using UnityEngine;
 using System.Collections;
 
-namespace ubv
+namespace ubv.common.data
 {
-    namespace common
+    public class GameStartMessage : serialization.Serializable
     {
-        namespace data
+        public serialization.types.List<PlayerState> Players;
+
+        protected override void InitSerializableMembers()
         {
-            public class GameStartMessage : udp.Serializable
-            {
-                public udp.SerializableTypes.List<common.data.PlayerState> Players;
+            Players = new serialization.types.List<PlayerState>(this, new System.Collections.Generic.List<PlayerState>());
+        }
 
-                protected override void InitSerializableMembers()
-                {
-                    Players = new udp.SerializableTypes.List<PlayerState>(this, new System.Collections.Generic.List<PlayerState>());
-                }
-
-                protected override byte SerializationID()
-                {
-                    return (byte)udp.Serialization.BYTE_TYPE.START_MESSAGE;
-                }
-            }
+        protected override byte SerializationID()
+        {
+            return (byte)serialization.ID.START_MESSAGE;
         }
     }
 }
