+++ Assets/Scripts/Common/ScriptableObjects.meta
