+++ Assets/Scripts/Serialization/Serialization.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d0aef6a8a84671946b14027cfc6808a5
+guid: 3b7e83fd4a5dc144a887daec80a27dc8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
